backend/requirements.txt
asgiref==3.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
